Add files: gyro_IICsoft and gyro_IIChard. Rebuild: gyroscpe.c
