+++ driver/src/STC32G_I2C.c
@@ -8,7 +8,7 @@
 /* --- Web: www.STCMCUDATA.com  ---------------------------------------*/
 /* --- BBS: www.STCAIMCU.com  -----------------------------------------*/
 /* --- QQ:  800003751 -------------------------------------------------*/
-/* 如果要在程序中使用此代码,请在程序中注明使用了STC的资料及程序            */
+/* 如果要在程序中使用此代码,请在程序中注明使用了STC的资料及程序        */
 /*---------------------------------------------------------------------*/
 
 #include	"STC32G_I2C.h"
