+++ user/include/control.h
@@ -5,7 +5,7 @@
 #include "motor.h"
 #include "encoder.h"
 #include "signal_acq.h"
-#include "gyroscope.h"
+#include "gyro_IIChard.h"
 
 void Update_Motors(PID_State * left_state,PID_State * right_state,PID_State * gyro_state);
 void Speed_Ctrl(unsigned int Left_Speed,unsigned int Right_Speed);
