+++ user/src/main.c
@@ -17,7 +17,6 @@ int main(void)
 	/******************************************************************/
 	EA =1;
 	/******************************************************************/
-	mpu6050_init();
 	//Ö÷Ñ­»·
 	while(1)
 	{
